+++ visualSHARK/util/helper.py
@@ -6,9 +6,203 @@ import math
 import networkx as nx
 import logging
 import timeit
+from datetime import datetime, timedelta
+
 from collections import deque
 
-from visualSHARK.models import Commit
+from mongoengine.queryset.visitor import Q
+
+from visualSHARK.models import Commit, FileAction, File, Hunk, CodeEntityState, VCSSystem, IssueSystem, Issue
+
+
+# TODO: split up date and message parsing
+class Label(object):
+
+    def __init__(self):
+        self._direct_link_jira = re.compile('(?P<ID>[A-Z][A-Z0-9_]+-[0-9]+)', re.M)
+        self._direct_link_bz = re.compile('(bug|issue|bugzilla)[s]{0,1}[#\s]*(?P<ID>[0-9]+)', re.I | re.M)
+        self._direct_link_gh = re.compile('(bug|issue|close|fixe[ds])[s]{0,1}[#\s]*(?P<ID>[0-9]+)', re.I | re.M)
+
+        # more in line with official documentation for auto-close
+        # https://help.github.com/articles/closing-issues-using-keywords/
+        self._direct_link_gh = re.compile('(resolve[ds]|bug|issue|close[ds]|fixe[ds])[s]{0,1}[#\s]*(?P<ID>[0-9]+)', re.I | re.M)
+
+        self._keyword = re.compile('(\s|^)fix(e[ds])?|(\s|^)bugs?|(\s|^)defects?|(\s|^)patch|(\s|^)issue[s]{0,1}', re.I | re.M)
+
+        self._numbers = re.compile('[#\s\,\-_/]+(?P<ID>[0-9]+)([#\s\,\-_$]|\.$|\.\s)', re.I | re.M)
+
+    def _clean_commit_message(self, msg):
+        ret = []
+        lines = msg.split('\n')
+
+        for line in lines:
+            # 1. remove git-svn-id line
+            if line.startswith('git-svn-id'):
+                continue
+            # 2. remove pull request line
+            if line.startswith('Merge pull request #'):
+                continue
+            ret.append(line)
+
+        return '\n'.join(ret)
+
+    def generate_candidates(self, commit):
+        msg = self._clean_commit_message(commit.message)
+
+        vcs = VCSSystem.objects.get(id=commit.vcs_system_id)
+        links = []
+        for its in IssueSystem.objects.filter(project_id=vcs.project_id):
+            if 'jira' in its.url:
+                links += self._get_links_jira(its, commit, msg)
+            elif 'github' in its.url:
+                links += self._get_links_github(its, commit, msg)
+
+        return links
+
+    def generate_affected_entities(self, commit, file_action_id):
+        entities = {}
+        for fa in FileAction.objects.filter(id=file_action_id):
+            file = File.objects.get(id=fa.file_id)
+
+            # we may have multiple hunks per file
+            for hunk in Hunk.objects.filter(file_action_id=fa.id):
+
+                # this prevents running out of bounds for entities
+                hunk_end = hunk.new_start + hunk.new_lines
+                if hunk.new_lines - hunk.old_lines > 0:
+                    hunk_end -= hunk.old_lines
+
+                # we want to have entities where either hunk start or hunk end is in their line range
+                q_objects = Q(start_line__lte=hunk.new_start, end_line__gte=hunk.new_start)  # hunk start is within entity
+                q_objects |= Q(start_line__lte=hunk_end, end_line__gte=hunk_end)  # hunk end is within entity
+                q_objects |= Q(start_line__lte=hunk.new_start, end_line__gte=hunk_end)  # hunk is completely within entity
+                q_objects |= Q(start_line__gte=hunk.new_start, end_line__lte=hunk_end)  # entity is within hunk
+
+                # print('file ', file.path)
+                # print('hunk start ', hunk.new_start, 'hunk_end', hunk_end)
+
+                # adopted for after memeshark run
+                # ces_ids = commit.code_entity_states
+                qry = CodeEntityState.objects.filter(file_id=file.id, id__in=commit.code_entity_states, ce_type__in=['method', 'class', 'interface'])  # we discard attribute and file
+                
+                # print('ces: ', len(qry))
+                qry = qry.filter(q_objects)
+                for ces in qry:
+                    k = str(ces.id)
+                    if k not in entities.keys():
+                        entities[k] = {'hunks': [{'hunk_id': str(hunk.id), 'hunk_start': hunk.new_start, 'hunk_end': hunk.new_start + hunk.new_lines - hunk.old_lines, 'hunk_content': hunk.content}], 'ces_id': str(ces.id), 'path': file.path, 'long_name': ces.long_name, 'type': ces.ce_type, 'start_line': ces.start_line, 'end_line': ces.end_line}
+                    else:
+                        entities[k]['hunks'].append({'hunk_id': str(hunk.id), 'hunk_start': hunk.new_start, 'hunk_end': hunk.new_start + hunk.new_lines - hunk.old_lines, 'hunk_content': hunk.content})
+
+            # files are a special case, they have no line_start and end so we add them according to the changed files of the commit
+            for ces in CodeEntityState.objects.filter(file_id=file.id, id__in=commit.code_entity_states, ce_type='file'):
+                k = str(ces.id)
+                if k not in entities.keys():
+                    entities[k] = {'hunks': [], 'ces_id': str(ces.id), 'path': file.path, 'long_name': ces.long_name, 'type': ces.ce_type, 'start_line': None, 'end_line': None}
+
+        ret = []
+        for k, v in entities.items():
+            ret.append(v)
+
+        return ret
+
+    def _get_commit_utc(self, commit):
+
+        # add/subtract offset to get utc
+        tmp = commit.committer_date.timestamp()
+        tmp += commit.committer_date_offset
+
+        # get datetime object from utc timestamp
+        return datetime.fromtimestamp(tmp)
+
+    def _get_links_github(self, its, commit, msg):
+        links = []
+        issue_ids = []
+
+        for m in self._direct_link_gh.finditer(msg):
+            issue_id = str(m.group('ID'))
+            issue_ids.append(issue_id)
+
+        for m in self._numbers.finditer(msg):
+            issue_id = str(m.group('ID'))
+            issue_ids.append(issue_id)
+
+        for issue_id in list(set(issue_ids)):
+            try:
+                # project_identifier = issue_id.split('-')[0]
+                # if project_identifier in ['ZK', 'KEEPER', 'ZOOKEEPR']:
+                #     issue_id = issue_id.replace(project_identifier, 'ZOOKEEPER')
+
+                i = Issue.objects.get(issue_system_id=its.id, external_id=issue_id)
+                tmp = {'issue_id': str(i.id), 'issue': issue_id, 'exists': True, 'type': i.issue_type, 'status': i.status, 'resolution': i.resolution, 'created_at': str(i.created_at), 'updated_at': str(i.updated_at), 'confidence': 1, 'confidence_reasons': []}
+                tmp['confidence_reasons'].append({'score': 1, 'reason': 'Issue with this ID found.'})
+
+                cd = self._get_commit_utc(commit)
+                if cd - timedelta(days=1) <= i.updated_at <= cd + timedelta(days=1):
+                    tmp['confidence'] += 1
+                    tmp['confidence_reasons'].append({'score': 1, 'reason': 'Issue last update date within one day of committer date.'})
+                else:
+                    td = cd - i.created_at
+                    tmp['confidence'] -= td.days / 365
+                    tmp['confidence_reasons'].append({'score': -td.days / 365, 'reason': 'Issue last update date not within one day of committer date.'})
+                links.append(tmp)
+
+                # if i.issue_type and i.issue_type.lower() == 'bug':
+                #     # print('{} status: {}, resolution: {}'.format(issue_id, i.status, i.resolution))
+                #     if i.status.lower() in ['resolved', 'closed'] and i.resolution.lower() in ['fixed']:
+                #         # print('found: {}'.format(issue_id))
+                #         defect_ids.append(issue_id)
+
+            # issue does not exist in our Database, still it is a candidate
+            except Issue.DoesNotExist as e:
+                tmp = {'issue_id': None, 'issue': issue_id, 'exists': False, 'type': None, 'status': None, 'resolution': None, 'created_at': None, 'updated_at': None, 'confidence': 0, 'confidence_reasons': []}
+                tmp['confidence'] -= 1
+                tmp['confidence_reasons'].append({'score': -1, 'reason': 'No issue with this ID.'})
+                links.append(tmp)
+
+        return links
+
+    def _get_links_jira(self, its, commit, msg):
+        links = []
+
+        issue_ids = []
+        for m in self._direct_link_jira.finditer(msg):
+            issue_id = str(m.group('ID'))
+            issue_ids.append(issue_id)
+
+        for issue_id in list(set(issue_ids)):
+            try:
+                # project_identifier = issue_id.split('-')[0]
+                # if project_identifier in ['ZK', 'KEEPER', 'ZOOKEEPR']:
+                #     issue_id = issue_id.replace(project_identifier, 'ZOOKEEPER')
+
+                i = Issue.objects.get(issue_system_id=its.id, external_id=issue_id)
+                tmp = {'issue_id': str(i.id), 'issue': issue_id, 'exists': True, 'type': i.issue_type, 'status': i.status, 'resolution': i.resolution, 'created_at': str(i.created_at), 'updated_at': str(i.updated_at), 'confidence': 1, 'confidence_reasons': []}
+                tmp['confidence_reasons'].append({'score': 1, 'reason': 'Issue with this ID found.'})
+
+                cd = self._get_commit_utc(commit)
+                if cd - timedelta(days=1) <= i.updated_at <= cd + timedelta(days=1):
+                    tmp['confidence'] += 1
+                    tmp['confidence_reasons'].append({'score': 1, 'reason': 'Issue last update date within one day of committer date.'})
+                else:
+                    td = cd - i.created_at
+                    tmp['confidence'] -= td.days / 365
+                    tmp['confidence_reasons'].append({'score': -td.days / 365, 'reason': 'Issue last update date not within one day of committer date.'})
+                links.append(tmp)
+                # if i.issue_type and i.issue_type.lower() == 'bug':
+                #     # print('{} status: {}, resolution: {}'.format(issue_id, i.status, i.resolution))
+                #     if i.status.lower() in ['resolved', 'closed'] and i.resolution.lower() in ['fixed']:
+                #         # print('found: {}'.format(issue_id))
+                #         defect_ids.append(issue_id)
+
+            # issue does not exist in our Database, still it is a candidate
+            except Issue.DoesNotExist as e:
+                tmp = {'issue_id': None, 'issue': issue_id, 'exists': False, 'type': None, 'status': None, 'resolution': None, 'created_at': None, 'updated_at': None, 'confidence': 0, 'confidence_reasons': []}
+                tmp['confidence'] -= 1
+                tmp['confidence_reasons'].append({'score': -1, 'reason': 'No issue with this ID.'})
+                links.append(tmp)
+
+        return links
 
 
 def tag_filter(tags, discard_qualifiers=True, discard_patch=False):