+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.1.2 (in progress)
+- Grid does now switch to first page if entries to display is changed
+- added first line of commit message to commit list views
+- added hunk api endpoint
+- added lines to code entity states endpoint
+- added issue link candidates generation based on ITS and commit message
+- added custom test runner for execution with mongomock (disregarding the MySQL Part for now)
+
 0.1.1
 - Successful jobs get acknowledged automatically
 - added Branch API Endpoint